Encrypt and store seed phrase
