+++ packages/browser-wallet/src/popup/pages/Setup/RecoveryPhrase.tsx
@@ -4,12 +4,13 @@ import { useNavigate } from 'react-router-dom';
 import { absoluteRoutes } from '@popup/constants/routes';
 import TextArea from '@popup/shared/Form/TextArea';
 import { passcodeAtom, seedPhraseAtom } from '@popup/state';
-import { useAtomValue } from 'jotai';
+import { useAtomValue, useSetAtom } from 'jotai';
 import Form from '@popup/shared/Form';
 import { SubmitHandler, Validate } from 'react-hook-form';
 import Submit from '@popup/shared/Form/Submit';
 import { useTranslation } from 'react-i18next';
-// import { encrypt } from '@popup/shared/crypto';
+import { encrypt } from '@popup/shared/crypto';
+import { encryptedSeedPhraseAtom } from '@popup/store/settings';
 import { setupRoutes } from './routes';
 
 type FormValues = {
@@ -20,6 +21,7 @@ export function EnterRecoveryPhrase() {
     const navigate = useNavigate();
     const { t } = useTranslation('setup');
     const seedPhrase = useAtomValue(seedPhraseAtom);
+    const setEncryptedSeedPhrase = useSetAtom(encryptedSeedPhraseAtom);
     const passcode = useAtomValue(passcodeAtom);
 
     if (!passcode) {
@@ -27,10 +29,9 @@ export function EnterRecoveryPhrase() {
         return null;
     }
 
-    const handleSubmit: SubmitHandler<FormValues> = () => {
-        // TODO Encrypt and store the recovery phrase here.
-
-        // console.log(JSON.stringify(encrypt(vs.seedPhraseInput, passcode)));
+    const handleSubmit: SubmitHandler<FormValues> = (vs) => {
+        const encryptedSeedPhrase = encrypt(vs.seedPhraseInput, passcode);
+        setEncryptedSeedPhrase(encryptedSeedPhrase);
         navigate(`${absoluteRoutes.setup.path}/${setupRoutes.chooseNetwork}`);
     };
