+++ packages/browser-wallet/src/popup/store/settings.ts
@@ -1,10 +1,15 @@
-import { ChromeStorageKey, Theme, WalletCredential } from '@shared/storage/types';
+import { ChromeStorageKey, EncryptedData, Theme, WalletCredential } from '@shared/storage/types';
 import { atom } from 'jotai';
 import { EventType } from '@concordium/browser-wallet-api-helpers';
 import { popupMessageHandler } from '@popup/shared/message-handler';
 import { atomWithChromeStorage, AsyncWrapper } from './utils';
 
 export const credentialsAtom = atomWithChromeStorage<WalletCredential[]>(ChromeStorageKey.Credentials, []);
+export const encryptedSeedPhraseAtom = atomWithChromeStorage<EncryptedData | undefined>(
+    ChromeStorageKey.SeedPhrase,
+    undefined,
+    true
+);
 export const themeAtom = atomWithChromeStorage<Theme>(ChromeStorageKey.Theme, Theme.Light);
 
 const storedJsonRpcUrlAtom = atomWithChromeStorage<string>(ChromeStorageKey.JsonRpcUrl, '', true);
