+++ packages/browser-wallet/src/shared/storage/types.ts
@@ -3,6 +3,7 @@ export enum ChromeStorageKey {
     Credentials = 'credentials',
     JsonRpcUrl = 'jsonRpcUrl',
     SelectedAccount = 'selectedAccont',
+    SeedPhrase = 'seedPhrase',
     Theme = 'theme',
 }
 
@@ -15,3 +16,18 @@ export type WalletCredential = {
     key: string;
     address: string;
 };
+
+interface EncryptionMetaData {
+    keyLen: number;
+    iterations: number;
+    salt: string;
+    initializationVector: string;
+    encryptionMethod: string;
+    keyDerivationMethod: string;
+    hashAlgorithm: string;
+}
+
+export interface EncryptedData {
+    cipherText: string;
+    metadata: EncryptionMetaData;
+}
